fixed action probability + changed import positions for cleaner imports
