+++ baselines/__init__.py
@@ -1,0 +1,9 @@
+from baselines.a2c import A2C
+from baselines.acer import ACER
+from baselines.acktr import ACKTR
+from baselines.ddpg import DDPG
+from baselines.deepq import DeepQ
+from baselines.gail import GAIL
+from baselines.ppo1 import PPO1
+from baselines.ppo2 import PPO2
+from baselines.trpo_mpi import TRPO
