+++ baselines/common/vec_env/base_vec_env.py
@@ -1,0 +1,155 @@
+from abc import ABC, abstractmethod
+import pickle
+
+import cloudpickle
+from baselines import logger
+
+
+class AlreadySteppingError(Exception):
+    """
+    Raised when an asynchronous step is running while
+    step_async() is called again.
+    """
+
+    def __init__(self):
+        msg = 'already running an async step'
+        Exception.__init__(self, msg)
+
+
+class NotSteppingError(Exception):
+    """
+    Raised when an asynchronous step is not running but
+    step_wait() is called.
+    """
+
+    def __init__(self):
+        msg = 'not running an async step'
+        Exception.__init__(self, msg)
+
+
+class VecEnv(ABC):
+    def __init__(self, num_envs, observation_space, action_space):
+        """
+        An abstract asynchronous, vectorized environment.
+
+        :param num_envs: (int) the number of environments
+        :param observation_space: (Gym Space) the observation space
+        :param action_space: (Gym Space) the action space
+        """
+        self.num_envs = num_envs
+        self.observation_space = observation_space
+        self.action_space = action_space
+
+    @abstractmethod
+    def reset(self):
+        """
+        Reset all the environments and return an array of
+        observations, or a tuple of observation arrays.
+
+        If step_async is still doing work, that work will
+        be cancelled and step_wait() should not be called
+        until step_async() is invoked again.
+
+        :return: ([int] or [float]) observation
+        """
+        pass
+
+    @abstractmethod
+    def step_async(self, actions):
+        """
+        Tell all the environments to start taking a step
+        with the given actions.
+        Call step_wait() to get the results of the step.
+
+        You should not call this if a step_async run is
+        already pending.
+        """
+        pass
+
+    @abstractmethod
+    def step_wait(self):
+        """
+        Wait for the step taken with step_async().
+
+        :return: ([int] or [float], [float], [bool], dict) observation, reward, done, information
+        """
+        pass
+
+    @abstractmethod
+    def close(self):
+        """
+        Clean up the environment's resources.
+        """
+        pass
+
+    def step(self, actions):
+        """
+        Step the environments with the given action
+
+        :param actions: ([int] or [float]) the action
+        :return: ([int] or [float], [float], [bool], dict) observation, reward, done, information
+        """
+        self.step_async(actions)
+        return self.step_wait()
+
+    def render(self, mode='human'):
+        """
+        Gym environment rendering
+
+        :param mode: (str) the rendering type
+        """
+        logger.warn('Render not defined for %s' % self)
+
+    @property
+    def unwrapped(self):
+        if isinstance(self, VecEnvWrapper):
+            return self.venv.unwrapped
+        else:
+            return self
+
+
+class VecEnvWrapper(VecEnv):
+    def __init__(self, venv, observation_space=None, action_space=None):
+        """
+        Vectorized environment base class
+
+        :param venv: (VecEnv) the vectorized environment to wrap
+        :param observation_space: (Gym Space) the observation space (can be None to load from venv)
+        :param action_space: (Gym Space) the action space (can be None to load from venv)
+        """
+        self.venv = venv
+        VecEnv.__init__(self, num_envs=venv.num_envs, observation_space=observation_space or venv.observation_space,
+                        action_space=action_space or venv.action_space)
+
+    def step_async(self, actions):
+        self.venv.step_async(actions)
+
+    @abstractmethod
+    def reset(self):
+        pass
+
+    @abstractmethod
+    def step_wait(self):
+        pass
+
+    def close(self):
+        return self.venv.close()
+
+    def render(self, mode='human'):
+        self.venv.render()
+
+
+class CloudpickleWrapper(object):
+    def __init__(self, var):
+        """
+        Uses cloudpickle to serialize contents (otherwise multiprocessing tries to use pickle)
+
+        :param var: (Any) the variable you wish to wrap for pickling with cloudpickle
+        """
+        self.var = var
+
+    def __getstate__(self):
+        return cloudpickle.dumps(self.var)
+
+    def __setstate__(self, obs):
+        self.var = pickle.loads(obs)
